+++ backend/ruoyi-system/src/main/java/com/ruoyi/system/domain/IcesManufacturePlan.java
@@ -64,11 +64,11 @@ public class IcesManufacturePlan extends BaseEntity {
     /**
      * 计划产品数量
      */
-    private Long mpQtyPlan;
+    private Float mpQtyPlan;
     /**
      * 已完成产品数量
      */
-    private Long mpQtyReal;
+    private Float mpQtyReal;
     /**
      * 已删除
      */
